drm/kms: add support for memory bandwidth considerations (v2)
Some GPUs (mostly in the lower server end of the market) have memory
bandwidth limitations that hit in the real world, however memory
bw calcs require bit depth to be worked out and at the moment, the
KMS interface doesn't know about bit depth until we set the mode.

This "overloads" the connector_type_id in the get connector ioctl
to pass a depth value to the kernel code. A new libdrm API is added
to utilise this. libdrm has always set this to 0 up until now.

This depth value is what the kms user proposed to use, the actual
API in libdrm is drmModeGetConnectorWithDepth and will follow
along.

This patch also adds RN50 support to the radeon kms to check
the memory bandwidth calcs and refuse the higher modes on these
chips.

v2: rebased onto latest kernel - fixup radeon usage.

Signed-off-by: Dave Airlie <airlied@redhat.com>