drivers/gpu/qxl: add new driver
QXL is a paravirtual graphics device used by the Spice virtual desktop
interface.

This version is not meant for kernel inclusion as many TODOs remain,
performance is sub par due to whole page allocation on the pci bar, and
missing off screen surfaces support.

It works with the modesetting X driver, and with an updated qxl X
driver.

Signed-off-by: Alon Levy <alevy@redhat.com>