drm/prime: keep a reference from the handle to exported dma-buf (v2)
This is likely the simplest solution to the problem, and seems
to work fine.

When we export a dma_buf from a gem object, we keep track of it
with the handle, we take a reference to the dma_buf. When we close
the handle (i.e. userspace is finished with the buffer), we drop
the reference to the dma_buf, and it gets collected.

I'd like to refrain from too much in this patch as I'd like it to
go to stable, so future cleanups should look at maybe reducing
the obj storing two pointers etc.

v1.1: move export symbol line back up.

v2: okay I had to do a bit more, as the first patch showed a leak
on one of my tests, that I found using the dma-buf debugfs support,
the problem case is exporting a buffer twice with the same handle,
we'd add another export handle for it unnecessarily, however
we now fail if we try to export the same object with a different gem handle,
however I'm not sure if that is a case I want to support, and I've
gotten the code to WARN_ON if we hit something like that.

Signed-off-by: Dave Airlie <airlied@redhat.com>