drm/fb_helper: move deferred fb checking into restore mode
On MST systems the monitors don't appear when we set the fb up,
but plymouth opens the drm device and holds it open while they
come up, when plymouth finishes and lastclose gets called we
don't do the delayed fb probe, so the monitor never appears on the
console.

Fix this by moving the delayed checking into the mode restore.

Signed-off-by: Dave Airlie <airlied@redhat.com>